+++ CHANGELOG.md
@@ -6,4 +6,5 @@ All notable changes to `plumb`, derived mechanically from the [monorepo commits]
 
 ### Changed
 
+- plumb: array-style run references ${o\[N\]\[K\]}; hero redesign ([#3593](https://github.com/indexable-inc/index/issues/3593)) ([`b66746b`](https://github.com/indexable-inc/index/commit/b66746ba8d381fd63cd1b1d38fd03c8bc25abb6c), 2026-07-18)
 - plumb: inspectable bash-subset shell, library-first ([#3578](https://github.com/indexable-inc/index/issues/3578)) ([#3579](https://github.com/indexable-inc/index/issues/3579)) ([`8aec425`](https://github.com/indexable-inc/index/commit/8aec4254530619fb00c9cec244694c1d84dc844b), 2026-07-18)